fix(restapi): close the page-overflow hole on /requests too
Same class as the /movies and /series fix: the requests page parameter bound
as uint32 with no maximum, and the offset arithmetic ((page-1)*limit in
uint32) wrapped — page=42949674 with limit=100 served rows 5..104 with a 200
instead of an empty page. The parameter is now uint16 with the bound declared
in the spec, so out-of-range values fail the bind with a 400.

+++ internal/server/restapi/gen.go
@@ -3432,7 +3432,7 @@ type RequestLimit = uint32
 type RequestMediaTypeParam string
 
 // RequestPage defines model for RequestPage.
-type RequestPage = uint32
+type RequestPage = uint16
 
 // RequestStatusParam defines model for RequestStatusParam.
 type RequestStatusParam string
